Initialize the database before starting the web host instead of using an IHostedService implementation
