+++ sandbox/OpenIddict.Sandbox.AspNetCore.Client/Program.cs
@@ -1,22 +1,24 @@
 ﻿using Microsoft.AspNetCore;
+using OpenIddict.Sandbox.AspNetCore.Client;
+using OpenIddict.Sandbox.AspNetCore.Client.Models;
 
-namespace OpenIddict.Sandbox.AspNetCore.Client;
-
-public static class Program
-{
 #if SUPPORTS_WEB_INTEGRATION_IN_GENERIC_HOST
-    public static void Main(string[] args) =>
-        CreateHostBuilder(args).Build().Run();
-
-    public static IHostBuilder CreateHostBuilder(string[] args) =>
-        Host.CreateDefaultBuilder(args)
-            .ConfigureWebHostDefaults(builder => builder.UseStartup<Startup>());
+var builder = Host.CreateDefaultBuilder(args);
+builder.ConfigureWebHostDefaults(builder => builder.UseStartup<Startup>());
 #else
-    public static void Main(string[] args) =>
-        CreateWebHostBuilder(args).Build().Run();
-
-    public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-        WebHost.CreateDefaultBuilder(args)
-               .UseStartup<Startup>();
+var builder = WebHost.CreateDefaultBuilder(args);
+builder.UseStartup<Startup>();
 #endif
+
+var app = builder.Build();
+
+// Before starting the host, create the database used to store the application data.
+//
+// Note: in a real world application, this step should be part of a setup script.
+await using (var scope = app.Services.CreateAsyncScope())
+{
+    var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
+    await context.Database.EnsureCreatedAsync();
 }
+
+await app.RunAsync();
