--- sandbox/OpenIddict.Sandbox.AspNetCore.Client/Worker.cs
@@ -1,21 +1,0 @@
-﻿using OpenIddict.Sandbox.AspNetCore.Client.Models;
-
-namespace OpenIddict.Sandbox.AspNetCore.Client;
-
-public class Worker : IHostedService
-{
-    private readonly IServiceProvider _provider;
-
-    public Worker(IServiceProvider provider)
-        => _provider = provider;
-
-    public async Task StartAsync(CancellationToken cancellationToken)
-    {
-        await using var scope = _provider.CreateAsyncScope();
-
-        var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
-        await context.Database.EnsureCreatedAsync(cancellationToken);
-    }
-
-    public Task StopAsync(CancellationToken cancellationToken) => Task.CompletedTask;
-}
