+++ sandbox/OpenIddict.Sandbox.WinForms.Client/Program.cs
@@ -1,8 +1,10 @@
+using System.Diagnostics;
 using Dapplo.Microsoft.Extensions.Hosting.WinForms;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+using Microsoft.Win32;
 using OpenIddict.Client;
 using OpenIddict.Sandbox.WinForms.Client;
 using static OpenIddict.Abstractions.OpenIddictConstants;
@@ -97,14 +99,43 @@ var host = new HostBuilder()
                        });
             });
 
-        // Register the worker responsible for creating the database used to store tokens
-        // and adding the registry entries required to register the custom URI scheme.
         //
         // Note: in a real world application, this step should be part of a setup script.
-        services.AddHostedService<Worker>();
+        // services.Insert(0, ServiceDescriptor.Singleton<IHostedService, Worker>());
     })
     .ConfigureWinForms<MainForm>()
     .UseWinFormsLifetime()
     .Build();
+
+// Before starting the host, create the database used to store the application data
+// and add the registry entries required to register the custom URI scheme.
+//
+// Note: in a real world application, this step should be part of a setup script.
+await using (var scope = host.Services.CreateAsyncScope())
+{
+    var context = scope.ServiceProvider.GetRequiredService<DbContext>();
+    await context.Database.EnsureCreatedAsync();
+
+    // Create the registry entries necessary to handle URI protocol activations.
+    //
+    // Note: this sample creates the entry under the current user account (as it doesn't
+    // require administrator rights), but the registration can also be added globally
+    // in HKEY_CLASSES_ROOT (in this case, it should be added by a dedicated installer).
+    //
+    // Alternatively, the application can be packaged and use windows.protocol to
+    // register the protocol handler/custom URI scheme with the operating system.
+    using var root = Registry.CurrentUser.CreateSubKey("SOFTWARE\\Classes\\com.openiddict.sandbox.winforms.client");
+    root.SetValue(string.Empty, "URL:com.openiddict.sandbox.winforms.client");
+    root.SetValue("URL Protocol", string.Empty);
+
+    using var command = root.CreateSubKey("shell\\open\\command");
+    command.SetValue(string.Empty, string.Format("\"{0}\" \"%1\"",
+#if SUPPORTS_ENVIRONMENT_PROCESS_PATH
+        Environment.ProcessPath
+#else
+        Process.GetCurrentProcess().MainModule.FileName
+#endif
+    ));
+}
 
 await host.RunAsync();
